+++ src/Modules/Region/RegionGateway.php
@@ -221,6 +221,12 @@ class RegionGateway extends BaseGateway
         ]);
     }
 
+    /**
+     * Fetches details for a region.
+     *
+     * Warning: this function does not properly set the moderated flag for large regions. In most cases you might want
+     * to use RegionTransactions::getRegionDetails instead.
+     */
     public function getRegionDetails(int $regionId): array
     {
         $region = $this->db->fetch('