+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Use Store DTO in StoreGateway !2515 !2611 @KoeMai
 - Move business logic for editing store into StoreTransaction !2515 @KoeMai
 - Fix TeamStoreStatus and move getStoreMarkers to StoreGateway !2560 @KoeMai
+- Reimplement create new forum thread in vue js !2627 @chriswalg
 
 
 ## Dev/Test/CI stuff